Refact: Alignment to design changes of MLPro 2 #15
