+++ test/howtos/oa/howto_oa_ca_006_run_kmeans_3d_dynamic.py
@@ -35,9 +35,7 @@ from mlpro.bf.streams.streams import *
 from mlpro.bf.streams.streams.clouds import *
 from mlpro.bf.various import Log
 from mlpro.oa.streams import *
-from mlpro_int_river.wrappers.clusteranalyzers import *
-
-
+from mlpro_int_river.wrappers.clusteranalyzers import WrRiverKMeans2MLPro
 
 
 
@@ -87,8 +85,6 @@ class Dynamic3DScenario(OAScenario):
 
 
 
-
-
 # 2 Prepare Demo/Unit test mode
 if __name__ == '__main__':
     cycle_limit = 2000
@@ -103,15 +99,12 @@ else:
 
 
 
-
 # 3 Instantiate the stream scenario
 myscenario = Dynamic3DScenario(
     p_mode=Mode.C_MODE_REAL,
     p_cycle_limit=cycle_limit,
     p_visualize=visualize,
-    p_logging=logging
-    )
-
+    p_logging=logging)
