+++ test/howtos/oa/howto_oa_ca_007_run_kmeans_3d_static_normalized.py
@@ -39,11 +39,8 @@ In particular you will learn:
 
 from mlpro.bf.streams.streams import *
 from mlpro.bf.various import Log
-
 from mlpro.oa.streams import *
-from mlpro_int_river.wrappers.clusteranalyzers import *
-
-
+from mlpro_int_river.wrappers.clusteranalyzers import WrRiverKMeans2MLPro
 
 
 
@@ -58,8 +55,6 @@ else:
     logging     = Log.C_LOG_NOTHING
     visualize   = False
     step_rate   = 1
-
-
 
 
 
@@ -132,16 +127,12 @@ class Static3DScenario(OAScenario):
 
 
 
-
-
 # 2 Instantiate the stream scenario
 myscenario = Static3DScenario(
     p_mode=Mode.C_MODE_REAL,
     p_cycle_limit=cycle_limit,
     p_visualize=visualize,
-    p_logging=logging
-    )
-
+    p_logging=logging)
