+++ test/howtos/oa/howto_oa_ca_012_run_streamkmeans_2d_dynamic.py
@@ -36,9 +36,7 @@ from mlpro.bf.streams.streams import *
 from mlpro.bf.streams.streams.clouds import *
 from mlpro.bf.various import Log
 from mlpro.oa.streams import *
-from mlpro_int_river.wrappers.clusteranalyzers import *
-
-
+from mlpro_int_river.wrappers.clusteranalyzers import WrRiverStreamKMeans2MLPro
 
 
 
@@ -87,7 +85,6 @@ class Dynamic2DScenario(OAScenario):
 
 
 
-
 # 2 Prepare Demo/Unit test mode
 if __name__ == '__main__':
     cycle_limit = 1000
@@ -102,15 +99,12 @@ else:
 
 
 
-
 # 3 Instantiate the stream scenario
 myscenario = Dynamic2DScenario(
     p_mode=Mode.C_MODE_REAL,
     p_cycle_limit=cycle_limit,
     p_visualize=visualize,
-    p_logging=logging
-    )
-
+    p_logging=logging)
