+++ test/howtos/oa/howto_oa_ca_013_run_streamkmeans_2d_static_normalized.py
@@ -41,9 +41,7 @@ from mlpro.bf.streams.streams import *
 from mlpro.bf.various import Log
 
 from mlpro.oa.streams import *
-from mlpro_int_river.wrappers.clusteranalyzers import *
-
-
+from mlpro_int_river.wrappers.clusteranalyzers import WrRiverStreamKMeans2MLPro
 
 
 
@@ -115,7 +113,6 @@ class Static2DScenario(OAScenario):
 
 
 
-
 # 2 Prepare Demo/Unit test mode
 if __name__ == '__main__':
     cycle_limit = 2000
@@ -130,15 +127,12 @@ else:
 
 
 
-
 # 3 Instantiate the stream scenario
 myscenario = Static2DScenario(
     p_mode=Mode.C_MODE_REAL,
     p_cycle_limit=cycle_limit,
     p_visualize=visualize,
-    p_logging=logging
-    )
-
+    p_logging=logging)
