+++ test/howtos/oa/howto_oa_ca_015_run_streamkmeans_3d_static.py
@@ -36,9 +36,7 @@ from mlpro.bf.streams.streams import *
 from mlpro.bf.streams.streams.clouds import *
 from mlpro.bf.various import Log
 from mlpro.oa.streams import *
-from mlpro_int_river.wrappers.clusteranalyzers import *
-
-
+from mlpro_int_river.wrappers.clusteranalyzers import WrRiverStreamKMeans2MLPro
 
 
 
@@ -88,8 +86,6 @@ class Static3DScenario(OAScenario):
 
 
 
-
-
 # 2 Prepare Demo/Unit test mode
 if __name__ == '__main__':
     cycle_limit = 1200
@@ -104,15 +100,12 @@ else:
 
 
 
-
 # 3 Instantiate the stream scenario
 myscenario = Static3DScenario(
     p_mode=Mode.C_MODE_REAL,
     p_cycle_limit=cycle_limit,
     p_visualize=visualize,
-    p_logging=logging
-    )
-
+    p_logging=logging)
