+++ test/howtos/oa/howto_oa_ca_016_run_streamkmeans_3d_dynamic.py
@@ -36,9 +36,7 @@ from mlpro.bf.streams.streams import *
 from mlpro.bf.streams.streams.clouds import *
 from mlpro.bf.various import Log
 from mlpro.oa.streams import *
-from mlpro_int_river.wrappers.clusteranalyzers import *
-
-
+from mlpro_int_river.wrappers.clusteranalyzers import WrRiverStreamKMeans2MLPro
 
 
 
@@ -89,8 +87,6 @@ class Dynamic3DScenario(OAScenario):
 
 
 
-
-
 # 2 Prepare Demo/Unit test mode
 if __name__ == '__main__':
     cycle_limit = 2000
@@ -105,15 +101,12 @@ else:
 
 
 
-
 # 3 Instantiate the stream scenario
 myscenario = Dynamic3DScenario(
     p_mode=Mode.C_MODE_REAL,
     p_cycle_limit=cycle_limit,
     p_visualize=visualize,
-    p_logging=logging
-    )
-
+    p_logging=logging)
