+++ test/howtos/oa/howto_oa_ca_022_run_clustream_2d_dynamic.py
@@ -33,9 +33,7 @@ from mlpro.bf.streams.streams import *
 from mlpro.bf.streams.streams.clouds import *
 from mlpro.bf.various import Log
 from mlpro.oa.streams import *
-from mlpro_int_river.wrappers.clusteranalyzers import *
-
-
+from mlpro_int_river.wrappers.clusteranalyzers import WrRiverCluStream2MLPro
 
 
 
@@ -89,7 +87,6 @@ class Dynamic2DScenario(OAScenario):
 
 
 
-
 # 2 Prepare Demo/Unit test mode
 if __name__ == '__main__':
     cycle_limit = 1000
@@ -104,15 +101,12 @@ else:
 
 
 
-
 # 3 Instantiate the stream scenario
 myscenario = Dynamic2DScenario(
     p_mode=Mode.C_MODE_REAL,
     p_cycle_limit=cycle_limit,
     p_visualize=visualize,
-    p_logging=logging
-    )
-
+    p_logging=logging)
