+++ test/howtos/oa/howto_oa_wr_002_river_clusteranalyzer_clustream.py
@@ -40,9 +40,7 @@ from mlpro.bf.streams.models import *
 from mlpro.bf.streams.streams.provider_mlpro import StreamMLProBase
 
 from mlpro.oa.streams import *
-from mlpro_int_river.wrappers.clusteranalyzers import *
-
-
+from mlpro_int_river.wrappers.clusteranalyzers import WrRiverCluStream2MLPro
 
 
 
@@ -135,11 +133,6 @@ class AdScenario4CluStream (OAScenario):
         return stream, workflow
 
 
-
-
-
-## -------------------------------------------------------------------------------------------------
-## -------------------------------------------------------------------------------------------------
 
 # 1 Preparation of demo/unit test mode
 if __name__ == "__main__":
