+++ test/howtos/oa/howto_oa_wr_003_river_clusteranalyzer_denstream.py
@@ -41,7 +41,7 @@ from mlpro.bf.streams.models import *
 from mlpro.bf.streams.streams.provider_mlpro import StreamMLProBase
 
 from mlpro.oa.streams import *
-from mlpro_int_river.wrappers.clusteranalyzers import *
+from mlpro_int_river.wrappers.clusteranalyzers.denstream import *
