+++ test/howtos/oa/howto_oa_wr_004_river_clusteranalyzer_kmeans.py
@@ -40,9 +40,7 @@ from mlpro.bf.streams.models import *
 from mlpro.bf.streams.streams.provider_mlpro import StreamMLProBase
 
 from mlpro.oa.streams import *
-from mlpro_int_river.wrappers.clusteranalyzers import *
-
-
+from mlpro_int_river.wrappers.clusteranalyzers import WrRiverKMeans2MLPro
 
 
 
@@ -133,10 +131,6 @@ class AdScenario4KMeans (OAScenario):
 
 
 
-
-
-## -------------------------------------------------------------------------------------------------
-## -------------------------------------------------------------------------------------------------
 
 # 1 Preparation of demo/unit test mode
 if __name__ == "__main__":
