+++ test/howtos/oa/howto_oa_wr_005_river_clusteranalyzer_streamkmeans.py
@@ -38,10 +38,8 @@ STREAMKMeans
 from mlpro.bf.streams.streams import *
 from mlpro.bf.streams.models import *
 from mlpro.bf.streams.streams.provider_mlpro import StreamMLProBase
-
 from mlpro.oa.streams import *
-from mlpro_int_river.wrappers.clusteranalyzers import *
-
+from mlpro_int_river.wrappers.clusteranalyzers import WrRiverStreamKMeans2MLPro
 
 
 
@@ -130,10 +128,6 @@ class AdScenario4STREAMKMeans (OAScenario):
 
 
 
-
-
-## -------------------------------------------------------------------------------------------------
-## -------------------------------------------------------------------------------------------------
 
 # 1 Preparation of demo/unit test mode
 if __name__ == "__main__":
